BACKEND/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
loguru==0.7.2
python-nmap==0.7.1
netifaces==0.11.0
fastapi-cache2==0.2.1
aiofiles==23.2.1
requests==2.25.1
